move resolve_protocols to cli, add docstring and test
